# manager/api/api-spec.bundled.yaml
openapi: 3.0.0
info:
  version: 0.0.0
  title: MaEVe CSMS
  description: Internal API to interact with the MaEVe CSMS, external clients should use OCPI.
  contact:
    name: MaEVe team
    email: maeve-team@thoughtworks.com
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:9410/api/v0
    description: The local development server
paths:
...
  /cs/{csId}/certificates:
    get:
      summary: Get installed certificates from the charge station
      description: |
        Queries the charge station for installed certificates. Uses OCPP GetInstalledCertificateIds
        to retrieve certificate hash data for certificates currently installed on the charge station.
      operationId: getInstalledCertificates
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
        - name: certificateType
          in: query
          description: Optional filter by certificate type
          required: false
          schema:
            type: string
            enum:
              - V2GRootCertificate
              - MORootCertificate
              - CSMSRootCertificate
              - V2GCertificateChain
              - ManufacturerRootCertificate
      responses:
        '202':
          description: Request accepted, will query charge station asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    post:
      summary: Install certificates on the charge station
      operationId: installChargeStationCertificates
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargeStationInstallCertificates'
      responses:
        '200':
          description: OK
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/certificate:
    delete:
      summary: Delete a certificate from the charge station
      description: |
        Requests the charge station to delete a specific certificate identified by its hash data.
        Uses OCPP DeleteCertificate to remove the certificate from the charge station.
      operationId: deleteChargeStationCertificate
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateHashDataRequest'
      responses:
        '202':
          description: Request accepted, will delete certificate asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/auth:
    get:
      summary: Returns the authentication details
      description: |
        Returns the details required by the CSMS gateway to determine how to authenticate
        the charge station
      operationId: lookupChargeStationAuth
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      responses:
        '200':
          description: Charge station auth response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargeStationAuth'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /certificate:
    post:
      summary: Upload a certificate
      description: |
        Uploads a client certificate to the CSMS. The CSMS can use the certificate to authenticate
        the charge station using mutual TLS when the TLS operations are being offloaded to a load-balancer.
      operationId: uploadCertificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Certificate'
      responses:
        '201':
          description: Created
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /certificate/{certificateHash}:
    get:
      summary: Lookup a certificate
      description: |
        Lookup a client certificate that has been uploaded to the CSMS using a base64 encoded SHA-256 hash
        of the DER bytes.
      operationId: lookupCertificate
      parameters:
        - required: true
          in: path
          name: certificateHash
          schema:
            type: string
            maxLength: 64
      responses:
        '200':
          description: Certificate details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      summary: Delete a certificate
      description: |
        Deletes a client certificate that has been uploaded to the CSMS using a base64 encoded SHA-256 hash
        of the DER bytes.
      operationId: deleteCertificate
      parameters:
        - required: true
          in: path
          name: certificateHash
          schema:
            type: string
            maxLength: 64
      responses:
        '204':
          description: No content
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /cs/{csId}/availability:
    post:
      summary: Change charge station availability
      description: |
        Changes the operational status of a charge station or specific connector.
        Used to take equipment out of service for maintenance or make it available again.
        Supports both OCPP 1.6 and OCPP 2.0.1.
      operationId: changeAvailability
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeAvailabilityRequest'
      responses:
        '202':
          description: Accepted - availability change request submitted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /cs/{csId}/reset:
    post:
      operationId: ResetChargeStation
      summary: Request a charge station reset
      parameters:
        - name: csId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetRequest'
      responses:
        '202':
          description: Reset request accepted
        '400':
          description: Invalid request
        '500':
          description: Internal server error
  /cs/{csId}/unlock-connector:
    post:
      operationId: UnlockConnector
      summary: Request to unlock a connector on a charge station
      parameters:
        - name: csId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockConnectorRequest'
      responses:
        '202':
          description: Unlock request accepted
        '400':
          description: Invalid request
        '500':
          description: Internal server error
...
components:
  schemas:
    ChargeStationAuth:
      type: object
      description: Connection details for a charge station
      required:
        - securityProfile
      properties:
        securityProfile:
          type: integer
          description: |
            The security profile to use for the charge station: * `0` - unsecured transport with basic auth * `1` - TLS with basic auth * `2` - TLS with client certificate
        base64SHA256Password:
          type: string
          maxLength: 64
          description: The base64 encoded, SHA-256 hash of the charge station password
        invalidUsernameAllowed:
          type: boolean
          description: If set to true then an invalid username will not prevent the charge station connecting
...
    ChargeStationInstallCertificates:
      type: object
      description: |
        The set of certificates to install on the charge station. The certificates will be sent
        to the charge station asynchronously.
      required:
        - certificates
      properties:
        certificates:
          type: array
          items:
            type: object
            description: The set of certificates to install
            required:
              - type
              - certificate
            properties:
              type:
                type: string
                enum:
                  - V2G
                  - MO
                  - MF
                  - CSMS
              certificate:
                type: string
                description: The PEM encoded certificate with newlines replaced by `\n`
              status:
                type: string
                description: The status, defaults to Pending
                enum:
                  - Accepted
                  - Rejected
                  - Pending
    CertificateHashDataRequest:
      type: object
      description: Certificate hash data for identifying a specific certificate
      required:
        - certificateHashData
      properties:
        certificateHashData:
          type: object
          required:
            - hashAlgorithm
            - issuerNameHash
            - issuerKeyHash
            - serialNumber
          properties:
            hashAlgorithm:
              type: string
              enum:
                - SHA256
                - SHA384
                - SHA512
              description: Algorithm used to generate the hash
            issuerNameHash:
              type: string
              description: Hash of the issuer's distinguished name
              maxLength: 128
            issuerKeyHash:
              type: string
              description: Hash of the issuer's public key
              maxLength: 128
            serialNumber:
              type: string
              description: Certificate serial number
              maxLength: 40
    OperationResponse:
      type: object
      description: Response for asynchronous operations
      properties:
        operationId:
          type: string
          description: Unique identifier for tracking the operation status
        status:
          type: string
          description: Initial operation status
          enum:
            - Pending
            - Accepted
...
    Status:
      type: object
      description: HTTP status
      required:
        - status
      properties:
        status:
          type: string
          description: The status description
        error:
          type: string
          description: The error details
    Certificate:
      type: object
      description: A client certificate
      required:
        - certificate
      properties:
        certificate:
          type: string
          description: The PEM encoded certificate with newlines replaced by `\n`
...
    ChangeAvailabilityRequest:
      type: object
      description: Request to change charge station or connector availability
      required:
        - type
      properties:
        connectorId:
          type: integer
          minimum: 0
          description: Connector ID (0 = entire station, >0 = specific connector)
          default: 0
        evseId:
          type: integer
          minimum: 0
          description: EVSE ID for OCPP 2.0.1 (optional)
        type:
          type: string
          enum:
            - Operative
            - Inoperative
          description: Target availability state
...
    ResetRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - Soft
            - Hard
          description: Type of reset to perform
    UnlockConnectorRequest:
      type: object
      required:
        - connectorId
      properties:
        connectorId:
          type: integer
          description: ID of the connector to unlock
...